spotless

+++ src/main/java/com/ytdd9527/networksexpansion/utils/ParticleUtil.java
@@ -94,7 +94,7 @@ public class ParticleUtil {
                 plugin.getServer().getScheduler().runTaskLaterAsynchronously(plugin, () -> finalRunnableList.forEach(Runnable::run), tick);
             }
 
-            time += interval;
+            time += (int) interval;
         }
     }
 
@@ -127,7 +127,7 @@ public class ParticleUtil {
                     }
                 }, time / 50);
             }
-            time += interval;
+            time += (int) interval;
         }
     }
